// src/app/core/api/model/movieDTO.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ActorDTO } from './actorDTO';


export interface MovieDTO { 
    id?: number;
    name: string;
    genre: MovieDTOGenreEnum;
    publicationDate: string;
    actors: Array<ActorDTO>;
}
export enum MovieDTOGenreEnum {
    Action = 'ACTION',
    Comedy = 'COMEDY',
    Drama = 'DRAMA',
    Horror = 'HORROR',
    Romance = 'ROMANCE',
    Sf = 'SF',
    Thriller = 'THRILLER',
    Animation = 'ANIMATION',
    Documentary = 'DOCUMENTARY'
};